cleanup and writeup
